removing duplicate hook and using bool to seperate hook user
